requirements.txt
absl-py==2.3.1
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dnspython==2.8.0
flatbuffers==25.9.23
fonttools==4.60.1
gitdb==4.0.12
gitpython==3.1.45
google-crc32c==1.7.1
idna==3.10
ifaddr==0.2.0
jax==0.7.2
jaxlib==0.7.2
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.6
mediapipe==0.10.14
ml-dtypes==0.5.3
narwhals==2.6.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt-einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==0.12.0
pyopenssl==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
streamlit==1.50.0
streamlit-webrtc==0.63.5
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
